### 2019-02-02 #1226 @whedon
[PRE REVIEW]: A Hyperbolic "Crystal Ball"
**Submitting author:** @benjaminwilson (<a href="http://orcid.org/0000-0002-3708-0361">Benjamin Wilson</a>)
**Repository:** <a href="https://github.com/lateral/crystal-ball" target ="_blank">https://github.com/lateral/crystal-ball</a>
**Version:** v1.0.1
