### 2019-02-02 #1226 @arfon
:wave: @benjaminwilson thanks for your submission to JOSS. From a quick inspection of this submission it's not entirely obvious that it meets our [submission criteria](http://joss.theoj.org/about#submission_requirements). In particular, this item:

> - Your software should have an obvious research application

Could you confirm here that there _is_ a research application for this software (and explain what that application is)? The section [_'what should my paper contain'_](http://joss.theoj.org/about#paper_structure) has some guidance for the sort of content we're looking to be present in the `paper.md`.

Many thanks!
