### 2019-02-04 007e5a7
Add RogueViz to related work section (fixes #1 ?)
