### 2019-02-04 d276f1a
paper.md updated to clarify research support as the core application
