+++ paper.md
@@ -17,7 +17,9 @@ bibliography: paper.bib
 
 # Summary
 
-Several recent works have presented methods for embeddings graphs [@chamberlain_2017] [@nickel_kiela_2017] [@NickelKiela2018] and words [@leimeister2018skipgram] [@Dhingra2018] in hyperbolic space.  Created for those experimenting with hyperbolic embeddings, the (Hyperbolic Crystal Ball)[https://github.com/lateral/crystal-ball] provides a mechanism for exploring these embeddings using the Poincaré disc model.  Consisting entirely of client-side JavaScript, the software can locally in any modern web browser.
+Several recent works have presented methods for embeddings graphs [@chamberlain_2017] [@nickel_kiela_2017] [@NickelKiela2018] and words [@leimeister2018skipgram] [@Dhingra2018] in hyperbolic space.
+The [Hyperbolic Crystal Ball](https://github.com/lateral/crystal-ball) provides researchers in machine learning with a much-needed mechanism for inspecting the hyperbolic embeddings resulting from their algorithms using a navigable visualisation on the Poincaré disc.
+Consisting entirely of client-side JavaScript, the software can locally in any modern web browser.
 
 The Poincaré disc displays the entirety of the hyperbolic plane within the unit disc.  In practice, however, due to the finite resolution of both the computer display and the eye, points that are far from the focal point of the disc are no longer visible.  Such point configurations occur frequently in hyperbolic embeddings, however.  The ``hyperbolic crystal ball`` is built to address this problem.  It provides a navigation mechanism for the hyperbolic plane, allowing the user to visit and explore far-off constellations of points.  This is achieved via simple dragging motions on the Poincaré disc.  These are able to effect arbitrary rotations and translations of the entire point configuration.
 
