### 2019-02-04 #1226 @benjaminwilson
Hi @arfon pleased to meet you.  The submission is in the category of "supports the functioning of research instruments or the execution of research experiments" (from the guidelines), rather than being research itself.  The core purpose is the inspection of hyperbolic embeddings resulting from machine learning pipelines, such as those cited in `paper.md` (the Leimeiser & Wilson paper is our own work).  I've tried to clarify this in the `paper.md`.
