### 2019-02-04 #1226 @poulson
Exploring large point clouds of embeddings is a standard step in research pipelines (especially if the embeddings can have a string associated with them). If this tool can visualize large amounts of points (e.g., at least thousands) in this manner, I can vouch that I would have used it in my last research job.
