### 2019-02-04 #1226 @benjaminwilson
Thanks @poulson . Indeed, the situation is compounded by the negative curvature of hyperbolic space, for which visualisations in the Euclidean plane, regardless of how they are derived (t-SNE, PCA, etc.) drastically distort the distances.

I could perhaps add some functionality for using custom labels - I'll add that to the list.

In terms of scale, I can report that I've tested in with 1.2k nodes and as many edges (the mammals graph from Nickel & Kiela, 2017).
