### 2019-02-04 #1226 @HaoZeke
@poulson I am a little backed up with another review but I would be glad to look this over next week if it is still relevant..
