### 2019-02-04 #1226 @poulson
Thank you for the quick comments @HaoZeke. I will wait to hear back from @Danbo3004 and, if they don't get back, I think by next week would be fine. It seems like you've already done some legwork.
