### 2019-02-07 7711b9f
Switch RogueViz author order, fixes #1
