### 2019-02-08 #1226 @HaoZeke
@poulson, I have completed my pending review, so I am free to look this over now.
