### 2019-02-08 #1240 @benjaminwilson
@HaoZeke thanks for your initial feedback over on the [other issue](https://github.com/openjournals/joss-reviews/issues/1226#issuecomment-460364318).  I'll address that today!
