### 2019-02-08 46cfe26
Expand paper.md on the maths; add performance example to README.md
