### 2019-02-08 f71168e
Small documentation improvements
