### 2019-02-08 #1240 @benjaminwilson
Hi @HaoZeke , I've added a section that gives a high-level view of the algorithms for computing the updates to the point configurations to `paper.md`, and added a small note about the performance to `README.md`.

I think this addresses the bulk of your points from previously - but not all of them - I'm afraid I'll have to leave those until Monday.

Thanks already for the comments to date!
