### 2019-02-09 lateral/crystal-ball#3 @HaoZeke
Add the version string to the code
Part of https://github.com/openjournals/joss-reviews/issues/1240. This is a very minor change, the version released and the on JOSS should be part of the source code. Something like:

```js
// Version 1.0.1
```
