### 2019-02-09 #1240 @HaoZeke
@benjaminwilson, thank you for the extensions, I will check when your additions are completed. In general, for the review process I will be opening issues on the project tracker, or where the solution is unambiguous, I shall be opening pull requests. Again, this is just the general outline, in this case I don't see much scope for either. Good work!
