+++ paper.bib
@@ -1,5 +1,3 @@
-
-
 @InProceedings{Sala_2018,
   title = 	 {Representation Tradeoffs for Hyperbolic Embeddings},
   author = 	 {Sala, Frederic and De Sa, Chris and Gu, Albert and Re, Christopher},
@@ -19,16 +17,19 @@
 }
 
 @ARTICLE{chamberlain_2017,
-  author = {{Chamberlain}, Benjamin~P. and {Clough}, James and {Deisenroth}, Marc~P.},
-  title = "{Neural Embeddings of Graphs in Hyperbolic Space}",
-  journal = {arXiv:1705.10359},
-  archivePrefix = "arXiv",
-  eprint = {1705.10359},
-  primaryClass = "stat.ML",
-  keywords = {Statistics - Machine Learning, Computer Science - Learning},
-  year = 2017,
-  url = {https://arxiv.org/abs/1705.10359},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+  author    = {Benjamin Paul Chamberlain and
+               James Clough and
+               Marc Peter Deisenroth},
+  title     = {Neural Embeddings of Graphs in Hyperbolic Space},
+  journal   = {CoRR},
+  volume    = {abs/1705.10359},
+  year      = {2017},
+  url       = {http://arxiv.org/abs/1705.10359},
+  archivePrefix = {arXiv},
+  eprint    = {1705.10359},
+  timestamp = {Mon, 13 Aug 2018 16:49:13 +0200},
+  biburl    = {https://dblp.org/rec/bib/journals/corr/ChamberlainCD17},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
 @misc{leimeister2018skipgram,
@@ -53,15 +54,14 @@
 }
 
 @article{Octavian2018,
-  author    = {Octavian{-}Eugen Ganea and
-               Gary B{\'{e}}cigneul and
-               Thomas Hofmann},
-  title     = {Hyperbolic Neural Networks},
-  journal   = {arxiv:1805.09112},
-  year      = {2018},
-  url       = {http://arxiv.org/abs/1805.09112},
-  archivePrefix = {arXiv},
-  eprint    = {1805.09112},
+title = {Hyperbolic Neural Networks},
+author = {Ganea, Octavian and Becigneul, Gary and Hofmann, Thomas},
+booktitle = {Advances in Neural Information Processing Systems 31},
+editor = {S. Bengio and H. Wallach and H. Larochelle and K. Grauman and N. Cesa-Bianchi and R. Garnett},
+pages = {5350--5360},
+year = {2018},
+publisher = {Curran Associates, Inc.},
+url = {http://papers.nips.cc/paper/7780-hyperbolic-neural-networks.pdf}
 }
 
 @InProceedings{Dhingra2018,
