### 2019-02-11 358e925
Add version number to all source files. Closes #3.
