+++ paper.md
@@ -19,19 +19,19 @@ bibliography: paper.bib
 
 Several recent works have presented methods for embeddings graphs [@chamberlain_2017] [@nickel_kiela_2017] [@NickelKiela2018] and words [@leimeister2018skipgram] [@Dhingra2018] in hyperbolic space.
 The [Hyperbolic Crystal Ball](https://github.com/lateral/crystal-ball) provides researchers in machine learning with a much-needed mechanism for inspecting the hyperbolic embeddings resulting from their algorithms using a navigable visualisation on the Poincaré disc.
-Consisting entirely of client-side JavaScript, the software can locally in any modern web browser.
+Consisting entirely of client-side JavaScript, the software is able to run locally in any modern web browser.
 
-The Poincaré disc displays the entirety of the hyperbolic plane within the unit disc.  In practice, however, due to the finite resolution of both the computer display and the eye, points that are far from the focal point of the disc are no longer visible.  Such point configurations occur frequently in hyperbolic embeddings, however.  The ``hyperbolic crystal ball`` is built to address this problem.  It provides a navigation mechanism for the hyperbolic plane, allowing the user to visit and explore far-off constellations of points.  This is achieved via simple dragging motions on the Poincaré disc.  These are able to effect arbitrary rotations and translations of the entire point configuration.
+The Poincaré disc displays the entirety of the hyperbolic plane within the familiar Euclidean unit disc.  In practice, however, due to the finite resolution of both the computer display and the eye, points that are far from the focal point of the disc are no longer visible.  Such point configurations occur frequently in hyperbolic embeddings, however.  The hyperbolic crystal ball is built to address this problem.  It provides a navigation mechanism for the hyperbolic plane, allowing the user to visit and explore far-off constellations of points.  This is achieved via simple dragging motions on the Poincaré disc.  These are able to effect arbitrary rotations and translations of the entire point configuration.
 
-In addition to this primary use case, the hyperbolic crystal ball is a useful tool for gaining an intuitive understanding of hyperbolic space and of the Poincaré disc model in particular.  To this end, a point-dragging functionality is included.  This permits the formation of arbitrary polygons, with edges being represented by geodesic arcs.  Viewing such polygons from various perspectives, by applying translations and rotations, provides an understanding of the deformations inherent in the Poincaré disc model.
+In addition to this primary use case, the hyperbolic crystal ball is a useful tool for gaining an intuitive understanding of hyperbolic space and of the Poincaré disc model in particular.  To this end, a point-dragging functionality is included.  This permits the formation of arbitrary polygons, with edges being represented by geodesic arcs.  Viewing such polygons from various perspectives (by applying translations and rotations) provides an understanding of the deformations inherent in the Poincaré disc model.
 
 # Computational geometry on the hyperbolic plane
 
-There are various models of the hyperbolic plane.  While the Poincaré disc model is used for display, the geometric computations of the crystal ball are carried out using the hyperboloid model.  The exception to this is the calculation of a geodesic line segment, which is acheived using circle inversion through the boundary of the Poincaré disc.
+There are various models of the hyperbolic plane.  While the Poincaré disc model is used for display, the geometric computations of the software are carried out using the hyperboloid model.  The exception to this is the calculation of a geodesic line segment, which is acheived using circle inversion through the boundary of the Poincaré disc.
 
 The point configuration can be manipulated either by dragging individual points or by dragging all points simultaneously.  The computations entailed involve the exponential, logarithm and geodesic parallel-transport maps on the hyperboloid.  The formulae for these maps are well-known, and can be found e.g. in [@leimeister2018skipgram].  Equipped with these maps, the user interaction and the consequent computations are as follows:
 
[…]
