### 2019-02-12 0523a48
Include sample data to show works with 1k nodes. Expand on precision.
