+++ sample-data/README.md
@@ -1,0 +1,8 @@
+# Sample data
+
+A balanced binary tree with 2 ** 10 = 1024 nodes (so 1024 - 1 = 1023 edges).  Derived using the combinatorial method of [Sala et al.](https://github.com/HazyResearch/hyperbolics).  To use this data, just copy/paste the JSON into the appropriate text fields.
+
+```
+binary-tree-edges.json
+binary-tree-points.json
+```
