### 2019-02-12 e5afd75
Hopefully fix missing DOIs!
