+++ paper.bib
@@ -16,20 +16,11 @@
   year = {2017}
 }
 
-@ARTICLE{chamberlain_2017,
-  author    = {Benjamin Paul Chamberlain and
-               James Clough and
-               Marc Peter Deisenroth},
-  title     = {Neural Embeddings of Graphs in Hyperbolic Space},
-  journal   = {CoRR},
-  volume    = {abs/1705.10359},
-  year      = {2017},
-  url       = {http://arxiv.org/abs/1705.10359},
-  archivePrefix = {arXiv},
-  eprint    = {1705.10359},
-  timestamp = {Mon, 13 Aug 2018 16:49:13 +0200},
-  biburl    = {https://dblp.org/rec/bib/journals/corr/ChamberlainCD17},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
+@inproceedings{chamberlain_2017,
+title={Neural Embeddings of Graphs in Hyperbolic Space},
+author={Ben Chamberlain, Marc Deisenroth and James Clough},
+booktitle={Proceedings of the 13th International Workshop on Mining and Learning with Graphs (MLG)},
+year={2017}
 }
 
 @misc{leimeister2018skipgram,
@@ -51,17 +42,6 @@
                10-15, 2018},
   pages     = {3776--3785},
   year      = {2018}
-}
-
-@article{Octavian2018,
-title = {Hyperbolic Neural Networks},
-author = {Ganea, Octavian and Becigneul, Gary and Hofmann, Thomas},
-booktitle = {Advances in Neural Information Processing Systems 31},
-editor = {S. Bengio and H. Wallach and H. Larochelle and K. Grauman and N. Cesa-Bianchi and R. Garnett},
-pages = {5350--5360},
-year = {2018},
-publisher = {Curran Associates, Inc.},
-url = {http://papers.nips.cc/paper/7780-hyperbolic-neural-networks.pdf}
 }
 
 @InProceedings{Dhingra2018,
