### 2019-02-12 #1240 @benjaminwilson
@HaoZeke thanks for your patience!  I think I've now addressed all issues raised in the pre-review.
