### 2019-02-12 #1240 @poulson
@benjaminwilson It seems that the DOIs were removed rather than fixed. Would you mind making sure that each of the references has a DOI?
