### 2019-02-13 5302b35
Tidy up the bibliography file.
