+++ paper.bib
@@ -1,59 +1,57 @@
-@InProceedings{Sala_2018,
-  title = 	 {Representation Tradeoffs for Hyperbolic Embeddings},
-  author = 	 {Sala, Frederic and De Sa, Chris and Gu, Albert and Re, Christopher},
-  booktitle = 	 {Proceedings of the 35th International Conference on Machine Learning},
-  pages = 	 {4460--4469},
-  year = 	 {2018},
-  address = 	 {Stockholmsmässan, Stockholm Sweden},
-  month = 	 {10--15 Jul},
-}
-
 @incollection{nickel_kiela_2017,
-  title = {Poincar\'{e} Embeddings for Learning Hierarchical Representations},
-  author = {Nickel, Maximillian and Kiela, Douwe},
-  booktitle = {Advances in Neural Information Processing Systems 30},
-  pages = {6338--6347},
-  year = {2017}
+	title = {Poincar\'{e} Embeddings for Learning Hierarchical Representations},
+	author = {Nickel, Maximillian and Kiela, Douwe},
+	booktitle = {Advances in Neural Information Processing Systems 30},
+	editor = {I. Guyon and U. V. Luxburg and S. Bengio and H. Wallach and R. Fergus and S. Vishwanathan and R. Garnett},
+	pages = {6338--6347},
+	year = {2017},
+	publisher = {Curran Associates, Inc.},
+	url = {http://papers.nips.cc/paper/7213-poincare-embeddings-for-learning-hierarchical-representations.pdf}
 }
 
 @inproceedings{chamberlain_2017,
-title={Neural Embeddings of Graphs in Hyperbolic Space},
-author={Ben Chamberlain, Marc Deisenroth and James Clough},
-booktitle={Proceedings of the 13th International Workshop on Mining and Learning with Graphs (MLG)},
-year={2017}
+	title={Neural Embeddings of Graphs in Hyperbolic Space},
+	author={Ben Chamberlain, Marc Deisenroth and James Clough},
+	booktitle={Proceedings of the 13th International Workshop on Mining and Learning with Graphs (MLG)},
+	year={2017},
+	url={http://www.mlgworkshop.org/2017/paper/MLG2017_paper_6.pdf}
 }
 
 @misc{leimeister2018skipgram,
-    title={Skip-gram word embeddings in hyperbolic space},
-    author={Matthias Leimeister and Benjamin J. Wilson},
-    year={2018},
-    eprint={1809.01498},
-    archivePrefix={arXiv},
-    primaryClass={cs.CL}
+	title={Skip-gram word embeddings in hyperbolic space},
+	author={Matthias Leimeister and Benjamin J. Wilson},
+	year={2018},
+	eprint={1809.01498},
+	archivePrefix={arXiv},
+	primaryClass={cs.CL}
 }
 
 @inproceedings{NickelKiela2018,
-  author    = {Maximilian Nickel and
-               Douwe Kiela},
-  title     = {Learning Continuous Hierarchies in the Lorentz Model of Hyperbolic
-               Geometry},
-  booktitle = {Proceedings of the 35th International Conference on Machine Learning,
-               {ICML} 2018, Stockholmsm{\"{a}}ssan, Stockholm, Sweden, July
-               10-15, 2018},
-  pages     = {3776--3785},
-  year      = {2018}
+	title =      {Learning Continuous Hierarchies in the {L}orentz Model of Hyperbolic Geometry},
+	author =      {Nickel, Maximillian and Kiela, Douwe},
+	booktitle =      {Proceedings of the 35th International Conference on Machine Learning},
+	pages =      {3779--3788},
+	year =      {2018},
[…]
