### 2019-02-13 #1240 @benjaminwilson
Hi @poulson, I've indeed removed some entries from the `.bib` file that I wasn't using.  For all but one of the remainder, I am afraid that I am really stumped on getting DOIs for any of them.  I was surprised that I couldn't find either on the paper itself or in published BibTeX or on crossref.org.  I don't know if this is because they are all either arXiv (1) or conference proceedings (rest).  Do you know of somewhere else I can look?
