### 2019-02-15 #1240 @poulson
@benjaminwilson Hi Benjamin, I am still relatively new to `whedon`'s `check references`, but you might try mirroring the DOI formatting of a recently completed review, such as https://github.com/sivaramambikasaran/HODLR/blob/master/docs/paper/paper.bib.

The DOI links are surrounded by curly braces and only contain the latter part of the address. Could you try this approach to see if the DOI's are picked up?
