### 2019-02-15 #1240 @benjaminwilson
Hi! The problem is rather that I can't find the DOIs for those articles anywhere.
