### 2019-02-15 6e93b6a
Update the sample data with more precise floats.
