### 2019-02-15 14a7696
Update arXiv citation without fuller .bib entry
