+++ paper.bib
@@ -17,13 +17,19 @@
 	url={http://www.mlgworkshop.org/2017/paper/MLG2017_paper_6.pdf}
 }
 
-@misc{leimeister2018skipgram,
-	title={Skip-gram word embeddings in hyperbolic space},
-	author={Matthias Leimeister and Benjamin J. Wilson},
-	year={2018},
-	eprint={1809.01498},
-	archivePrefix={arXiv},
-	primaryClass={cs.CL}
+@article{leimeister2018skipgram,
+	author    = {Matthias Leimeister and
+	             Benjamin J. Wilson},
+	title     = {Skip-gram word embeddings in hyperbolic space},
+	journal   = {CoRR},
+	volume    = {abs/1809.01498},
+	year      = {2018},
+	url       = {http://arxiv.org/abs/1809.01498},
+	archivePrefix = {arXiv},
+	eprint    = {1809.01498},
+	timestamp = {Fri, 05 Oct 2018 11:34:52 +0200},
+	biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1809-01498},
+	bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
 @inproceedings{NickelKiela2018,
