### 2019-02-20 #1240 @arfon
> @arfon This paper should be ready for submission; I am forwarding it over to you (please excuse me, I think there is a rotating editorial board but am not sure who that lands on right now).

@poulson - yep it's me this week. In the future you can ping @openjournals/joss-eics to get the editor on rotation.
